Split out router setup
